Move principal rendering to its own module
