+++ frontend/src/app/modules/common/openproject-common.module.ts
@@ -79,7 +79,6 @@ import {AutofocusDirective} from './autofocus/autofocus.directive';
 import {ShowSectionDropdownComponent} from './hide-section/show-section-dropdown.component';
 import {SlideToggleComponent} from './slide-toggle/slide-toggle.component';
 import {DynamicBootstrapModule} from './dynamic-bootstrap/dynamic-bootstrap.module';
-import {OpPrincipalComponent} from './principal/op-principal.component';
 import {OpFormFieldComponent} from './form-field/form-field.component';
 import {OpFormBindingDirective} from './form-field/form-binding.directive';
 import {OpOptionListComponent} from './option-list/option-list.component';
@@ -187,7 +186,6 @@ export function bootstrapModule(injector:Injector) {
     // filter
 
     SlideToggleComponent,
-    OpPrincipalComponent,
 
     OpFormFieldComponent,
     OpFormBindingDirective,
@@ -237,7 +235,6 @@ export function bootstrapModule(injector:Injector) {
 
     // User Avatar
     UserAvatarComponent,
-    OpPrincipalComponent,
 
     PersistentToggleComponent,
     HideSectionLinkComponent,
