+++ frontend/src/app/modules/principal/principal.component.ts
@@ -31,9 +31,10 @@ import {I18nService} from 'core-app/modules/common/i18n/i18n.service';
 import {PathHelperService} from 'core-app/modules/common/path-helper/path-helper.service';
 import {TimezoneService} from 'core-components/datetime/timezone.service';
 import {APIV3Service} from "core-app/modules/apiv3/api-v3.service";
-import {PrincipalHelper} from "core-app/modules/common/principal/principal-helper";
+
+import {PrincipalHelper} from "./principal-helper";
 import PrincipalPluralType = PrincipalHelper.PrincipalPluralType;
-import {PrincipalLike, PrincipalRendererService} from "core-app/modules/common/principal/principal-renderer.service";
+import {PrincipalLike, PrincipalRendererService} from "./principal-renderer.service";
 
 @Component({
   template: '',
